78595- Added helper class to download the exisitng EES event data AB#78595 AB#79953
