+++ src/UKHO.ERPFacade.API/Services/IERPFacadeService.cs
@@ -1,0 +1,9 @@
+﻿using UKHO.ERPFacade.API.Models;
+
+namespace UKHO.ERPFacade.API.Services
+{
+    public interface IERPFacadeService
+    {
+        List<UnitsOfSalePrices> BuildUnitOfSalePricePayload(List<PriceInformationEvent> priceInformationList);
+    }
+}
