+++ src/UKHO.ERPFacade.Common/IO/Azure/IAzureBlobEventWriter.cs
@@ -5,5 +5,7 @@
         Task UploadEvent(string requestEvent, string blobContainerName, string blobName);
 
         bool CheckIfContainerExists(string containerName);
+
+        string DownloadEvent(string blobName, string blobContainerName);
     }
 }
